requirements.txt
numpy
pandas
pytest
pytest-cov
pytest-qgis
scipy
shapely
geopandas
matplotlib
pydantic
python-docx
deep-translator
requests
